Module dioxus_fullstack::prelude

A prelude of commonly used items in dioxus-fullstack.

Re-exports

pub use server_fn::ServerFn as _;

pub use server_fn;

Enums

ServerFnError

Type for errors that can occur when using server functions.

Traits

DioxusServerFn

Defines a “server function.” A server function can be called from the server or the client, but the body of its code will only be run on the server, i.e., if a crate feature ssr is enabled.

Functions

get_root_props_from_document

Get the props from the document. This is only available in the browser.

server_cached

This allows you to send data from the server to the client. The data is serialized into the HTML on the server and hydrated on the client.

use_server_future

A future that resolves to a value.

Attribute Macros

server

Declares that a function is a server function. This means that its body will only run on the server, i.e., when the ssr feature is enabled.