Trait dioxus_fullstack::prelude::DioxusServerFn

pub trait DioxusServerFn: ServerFn<()> {
    // Provided method
    fn register_explicit() -> Result<(), ServerFnError> { ... }
}

Defines a “server function.” A server function can be called from the server or the client, but the body of its code will only be run on the server, i.e., if a crate feature ssr is enabled.

Server functions are created using the server macro.

The set of server functions can be queried on the server for routing purposes by calling [server_fn::ServerFunctionRegistry::get].

Technically, the trait is implemented on a type that describes the server function’s arguments, not the function itself.

Provided Methods

fn register_explicit() -> Result<(), ServerFnError>

Registers the server function, allowing the client to query it by URL.

Object Safety

This trait is not object safe.